% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anacuali.R
\name{anacuali}
\alias{anacuali}
\title{Analisis qualitativa do IMC corporal}
\usage{
anacuali(x)
}
\arguments{
\item{x}{Valor do IMC corporal}
}
\value{
Analisis qualitativa do IMC em funcao do limite = 22
}
\description{
Analisis qualitativa do IMC corporal
}
\examples{
medidas <- data.frame(paciente = seq(1:5), peso = c(65.8,48.3,59.6,63.2,51.9),
altura = c(1.65, 1.50, 1.59, 1.58, 1.61))

medidas

medidas$imc <- myimcf(medidas$peso, medidas$altura)

medidas$index <- anacuali(medidas$imc)

}
