% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesoptm.R
\name{pesoptm}
\alias{pesoptm}
\title{Sugestao peso ideal}
\usage{
pesoptm(peso, imc, altura)
}
\arguments{
\item{peso}{Calculado em Kg}

\item{imc}{Calculado em funcao do peso e altura}

\item{altura}{Calculado em m}
}
\value{
Sugestao do peso ideal em funcao de imc=22
}
\description{
Sugestao peso ideal
}
\examples{
medidas <- data.frame(paciente = seq(1:5), peso = c(65.8,48.3,59.6,63.2,51.9),
altura = c(1.65, 1.50, 1.59, 1.58, 1.61))
medidas$imc <- myimcf(medidas$peso, medidas$altura)
medidas$index <- anacuali(medidas$imc)

medidas$PesoIdeal <- pesoptm(medidas$peso, medidas$imc, medidas$altura)
medidas

}
